requirements.txt
beautifulsoup4==4.11.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
construct==2.8.8
cssutils==2.6.0
ffmpeg==1.4
future==0.18.3
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-core==2.3.2
google-cloud-speech==2.17.3
google-cloud-translate==3.10.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
iso8601==1.1.0
isodate==0.6.1
lxml==4.9.2
m3u8==3.4.0
mpegdash==0.3.1
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycaption==2.1.1
pycryptodome==3.17
pymp4==1.2.0
requests==2.28.2
rev-ai==2.17.1
rsa==4.9
six==1.16.0
soupsieve==2.4
urllib3==1.26.14
websocket-client==0.59.0
